fix: add prepaid credits from step modal
